src/test/java/io/okdp/spark/authc/model/AuthStateTest.java
/*
 *    Copyright 2024 tosit.io
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package io.okdp.spark.authc.model;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

import org.junit.jupiter.api.Test;

public class AuthStateTest {

  @Test
  public void test() {

    // When
    AuthState authState = AuthState.randomState();

    // Then
    assertThat(authState.state()).isNotBlank();
    assertThat(authState.codeVerifier().length()).isBetween(43, 128);
    assertThat(authState.codeVerifier()).isASCII();
  }
}
